🐞👴 Checkpoint ./model/action_queue.py:9665611/1025 ./utils/post.py:9665611/84

+++ utils/post.py
@@ -82,6 +82,8 @@ class Post:
     self._sendRequest(postData)
     
   def replacePost(self, message):
+    if not self._isAllowedToPostInThisChannel(self._event.channel()):
+      return;
     identity = IDENTITIES[self._event.user()]
     self.addMessage(message, Identity(identity.get('username'), identity.get('profilePicture')))
     self._deleteMessage()
